fix(tasks): quality audit triton2triton/geak_eval/L1/refk_identity
